requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
future==0.18.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow
platformdirs==2.6.0
pyserial==3.5
python-barcode==0.14.0
python-escpos==3.0a8
pyusb==1.2.1
PyYAML
qrcode==7.3.1
six==1.16.0
tomli==2.0.1
viivakoodi==0.8.0
Werkzeug==2.2.2
gunicorn
